(fu) > dropzone wrapper: handles pb media block - (file) add, sort and remove w/o ajax

+++ src/Presentation/SmartStore.Web/Scripts/smartstore.dropzoneWrapper.js
@@ -495,15 +495,55 @@
 						}
 					});
 				} else {
+
 					const hidden = document.getElementById("media-file-ids");
-					if (hidden) {
+					if (hidden && clearAssignableFiles) {
 						hidden.value = (hidden.value + assignableFileIds);
+
+						const items = previewContainer.find('.dz-image-preview');
+						const hiddenValues = hidden.value.split(',').map(Number);
+
+						$.each(items, function (index, item) {
+							if (index < items.length - assignableFiles.length) return;
+
+							const fileName = $(item).find(".fu-file-info-name").text();
+							var file = assignableFiles.find(x => x.media.id === hiddenValues[index]);
+							if (!file) {							
+								file = assignableFiles.find(x => x.name.toLowerCase() === fileName.toLowerCase());
+							}
+
+							if (file) {
+								// Set properties for newly added file preview.
+								var elPreview = file.previewElement ? $(file.previewElement) : $(previewTemplate.html());
+
+								elPreview
+									.attr("data-media-id", hiddenValues[index])
+									.attr("data-media-name", fileName)
+									.removeClass("d-none dz-processing");
+
+								elPreview.find(".fu-file-info-name").html(fileName);
+
+								elPreview
+									.find('img')
+									.attr('src', file.dataURL || file.media.thumbUrl);
+
+								previewContainer.append(elPreview);
+								dzResetProgressBar(elPreview.find(".progress-bar"));
+							}
+							else {
+								console.log("Error while adding preview element.", fileName.toLowerCase());
+							}
+						});
+
+						assignableFileIds = "";
+						assignableFiles.length = 0;
                     }
                 }
 			}
 
 			// Calls server function after sorting, to save current sort order.
 			function sortMediaFiles() {
+
 				if ($el.data('sort-url') && $el.data('entity-id')) {
 					var items = previewContainer.find('.dz-image-preview');
 
@@ -537,6 +577,19 @@
 							});
 						}
 					});
+				}
+				else {
+					const hidden = document.getElementById("media-file-ids");
+					if (hidden) {
+						const items = previewContainer.find('.dz-image-preview');
+
+						let newOrder = [];
+						$.each(items, function (i, val) {
+							newOrder.push($(val).data('entity-media-id'));
+						});
+
+						hidden.value = newOrder;
+					}
 				}
 			}
 
@@ -587,22 +640,43 @@
 				var entityMediaFileId = previewThumb.data("entity-media-id");
 				var mediaFileId = previewThumb.data("media-id");
 
-				$.ajax({
-					async: false,
-					cache: false,
-					type: 'POST',
-					url: $el.data('remove-url'),
-					data: { id: entityMediaFileId },
-					success: function () {
+				if ($el.data('remove-url')) {
+
+					$.ajax({
+						async: false,
+						cache: false,
+						type: 'POST',
+						url: $el.data('remove-url'),
+						data: { id: entityMediaFileId },
+						success: function () {
+							previewThumb.remove();
+
+							// File must be removed from dropzone if it was added in current queue.
+							var file = el.files.find(file => file.media.id === mediaFileId);
+							if (file)
+								el.removeFile(file);
+						}
+					});
+				} else {
+					const hidden = document.getElementById("media-file-ids");
+					if (hidden) {
 						previewThumb.remove();
 
 						// File must be removed from dropzone if it was added in current queue.
-						var file = el.files.find(file => file.media.id === mediaFileId);
+						const file = el.files.find(file => file.media.id === mediaFileId);						
 						if (file)
 							el.removeFile(file);
-					}
-				});
 
+						const items = previewContainer.find('.dz-image-preview');
+
+						let newOrder = [];
+						$.each(items, function (i, val) {
+							newOrder.push($(val).data('entity-media-id'));
+						});
+
+						hidden.value = newOrder;
+					}
+                }
 				return false;
 			});
